pyproject.toml
[project]
name = "poly-hammer-docs"
version = "1.0.0"
description = "Poly Hammer documentation site"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.2.2",
]

[dependency-groups]
dev = [
    "mkdocs>=1.6",
    "mkdocs-shadcn>=0.10",
    "pymdown-extensions>=10.0",
    "pygments>=2.18",
    "pyyaml>=6.0",
    "pre-commit>=4.0",
]

[tool.poly-hammer-docs]
github_org = "poly-hammer"
repos = [
    "meta-human-dna-addon",
    "character-control-rig-addon"
]
